Updated Documentation and added more app features.
The feature to generate random passwords is now available in the modal when entering new information. The user can now copy the passwords to their clipboard. (This feature required HTTPS and the application uses a dummy certificate). Minor changes to the UI and the removal of outdated files were also made.